requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-dotenv